+++ lib/raxx/http1.ex
@@ -723,7 +723,6 @@ defmodule Raxx.HTTP1 do
     payload_headers =
       case content_length(headers) do
         nil ->
-          # NOTE `:erlang.iolist_size/1` accepts binaries, i.e. should be `:erlang.iodata_size/1`
           case :erlang.iolist_size(iodata) do
             0 ->
               []